Use xmalloc, xcalloc and xrealloc

+++ ui/print_protocol.h
@@ -13,7 +13,7 @@ struct dns_info;
 struct nbns_info;
 
 /* write packet to buffer */
-void write_to_buf(char *buf, int size, struct packet *p);
+void write_to_buf(char *buf, size_t size, struct packet *p);
 
 void print_dns_record(struct dns_info *info, int i, char *buf, int n, uint16_t type);
 void print_nbns_record(struct nbns_info *info, int i, char *buf, int n);
